Move kas-core::root to kas-core::window and make pub; move Decorations there
