+++ crates/kas-core/src/event/cx/cx_pub.rs
@@ -18,7 +18,7 @@ use crate::theme::SizeCx;
 #[cfg(all(wayland_platform, feature = "clipboard"))]
 use crate::util::warn_about_error;
 #[allow(unused)] use crate::{Events, Layout, Tile}; // for doc-links
-use crate::{HasId, PopupDescriptor, Window};
+use crate::{HasId, PopupDescriptor, window::Window};
 
 impl EventState {
     /// Get the platform
