+++ crates/kas-core/src/event/cx/press/touch.rs
@@ -9,7 +9,8 @@ use super::{GrabMode, Press, PressSource, velocity};
 use crate::config::EventWindowConfig;
 use crate::event::{Event, EventCx, EventState, FocusSource};
 use crate::geom::{Affine, Coord, DVec2, Vec2};
-use crate::{Action, Id, NavAdvance, Node, Widget, Window};
+use crate::window::Window;
+use crate::{Action, Id, NavAdvance, Node, Widget};
 use cast::{Cast, CastApprox, Conv};
 use smallvec::SmallVec;
 use winit::event::TouchPhase;
