+++ crates/kas-core/src/popup.rs
@@ -8,7 +8,8 @@
 use crate::dir::Direction;
 use crate::event::{ConfigCx, Event, EventCx, IsUsed, Scroll, Unused, Used};
 use crate::layout::Align;
-use crate::{ChildIndices, Events, Id, Tile, Widget, WindowId};
+use crate::window::WindowId;
+use crate::{ChildIndices, Events, Id, Tile, Widget};
 use kas_macros::{impl_self, widget_index};
 
 #[allow(unused)] use crate::event::EventState;
