+++ crates/kas-core/src/prelude.rs
@@ -20,12 +20,11 @@ pub use crate::geom::{Coord, Offset, Rect, Size};
 pub use crate::layout::{Align, AlignHints, AlignPair, AxisInfo, SizeRules, Stretch};
 #[doc(no_inline)] pub use crate::text::AccessString;
 #[doc(no_inline)] pub use crate::theme::{DrawCx, SizeCx};
+#[doc(no_inline)] pub use crate::window::{Window, WindowId};
 #[doc(no_inline)]
 pub use crate::{ChildIndices, Node, Scrollable};
 #[doc(no_inline)]
-pub use crate::{
-    Events, Layout, Role, RoleCx, RoleCxExt, Tile, TileExt, Widget, Window, WindowCommand,
-};
+pub use crate::{Events, Layout, Role, RoleCx, RoleCxExt, Tile, TileExt, Widget};
 #[doc(no_inline)] pub use crate::{HasId, Id};
 #[doc(no_inline)]
 pub use crate::{autoimpl, impl_anon, impl_default, impl_scope, impl_self};
