+++ crates/kas-widgets/src/lib.rs
@@ -8,7 +8,6 @@
 //! ## Complex widgets
 //!
 //! -   [`EventConfig`] provides an editor for event configuration
-//! -   [`kas::Window`] is the root of any UI tree used as a window
 //! -   [`kas::Popup`] is the root of any popup
 //!
 //! ## Sub-modules
