+++ src/runner.rs
@@ -8,11 +8,10 @@
 //! Start by constructing a [`Runner`] or its [`Default`](type@Default)
 //! type-def (requires a backend be enabled, e.g. "wgpu").
 
-use crate::Window;
-use crate::WindowId;
 use crate::config::{AutoFactory, Config, ConfigFactory};
 use crate::draw::DrawSharedImpl;
 use crate::theme::Theme;
+use crate::window::{Window, WindowId};
 pub use kas_core::runner::{AppData, ClosedError, Error, MessageStack, Platform, Proxy, Result};
 use kas_core::runner::{GraphicsInstance, PreLaunchState};
 use kas_core::theme::FlatTheme;
